Create README.md
#Discription
Distributed Operation Layer : 
The distributed operation layer (DOL) is a software development framework to program parallel applications. The DOL allows to specify applications based on the Kahn process network model of computation and features a simulation engine based on SystemC. Moreover, the DOL provides an XML-based specification format to describe the implementation of a parallel application on a multi-processor systems, including binding and mapping.
#How to install
I didn't install the ubuntu, I simply copy the iso file given by TA Zhuoqun Guan
#Experimental experience
A good computer is crucial to a software engineer, and the system Windows 10 is the worst product ever made.